Task#hierfiles and Task#hierdir?
Coming up for a name for a "file that is nested in a directory that is
not meant to be interpreted as a node for linking or copying" is kind of
difficult.

Probably isn't appropriate given the German `hier`, which actually
implies the opposite of what we want.

Task#nicht_hier_files perhaps.