Revert "Haus::Link creates relative links by default"
This reverts commit 21a6790e9b26288d07ccba6bfd79c8c712c1c56f.

I forgot about the circumstance in which a link is on a symlinked node;
in that case, programs will either choose to follow the `physical` path,
or to happily traverse links. In this case, there are two right answers!

We'll have to add a `--physical` option to make the `--relative` option
viable.

Conflicts:

	README.markdown